Move GetRecordingStreamMediaSources to IMediaSourceManager
